fix: handle large Qwen3.5 vision batches safely
Use 64-bit offsets in the fused vision RoPE kernel. Bound independent patch projections and LayerNorm calls below signed 32-bit element limits to avoid large-batch indexing and numerical errors while preserving parameter names.

+++ rtp_llm/multimodal/multimodal_mixins/qwen3_5_moe/vision_kernels.py
@@ -86,7 +86,9 @@ def _rope_kernel(
     S1: tl.constexpr,
     BLOCK: tl.constexpr,
 ):
-    i = tl.program_id(0) * BLOCK + tl.arange(0, BLOCK)
+    # Packed QKV strides and the flattened output can exceed signed int32.
+    # Widen before multiplying so both input and output offsets stay valid.
+    i = tl.program_id(0).to(tl.int64) * BLOCK + tl.arange(0, BLOCK)
     mask = i < elements
     token = i // (HEADS * DIM)
     head = (i // DIM) % HEADS
